MidasSmall map complete.
Good little peasant units collecting all the stuff! They make me so proud.

It was touch and go there for a while but they negotiated for a higher cut
of the gold and in return they promised to walk faster. I think the
arrangment has worked out well for all.