BOTH build peasant xml files given were wrong.
Fixed them.

The small build peasants needed to collect 1000 wood and 1000
gold and it was set to 700 so the sequence wasn't ending.

The large build peasants asked for 3000 wood and 2000 gold and there is
only 2000 wood on the map so we think that the values should be switched.

Other small name changes and clean up are included in this commit.